+++ limesium-reactive-logging/src/test/kotlin/eu/inqudium/limesium/reactive/logging/CoRequestLoggingWebFilterCoroutineIntegrationTest.kt
@@ -4,6 +4,7 @@ import ch.qos.logback.classic.Level
 import ch.qos.logback.classic.Logger
 import ch.qos.logback.classic.spi.ILoggingEvent
 import ch.qos.logback.classic.spi.IThrowableProxy
+import eu.inqudium.limesium.common.AwaitingAppender
 import eu.inqudium.limesium.common.CorrelationIdGenerator
 import eu.inqudium.limesium.common.MdcKeys
 import eu.inqudium.limesium.common.NanoTimeSource