+++ limesium-reactive-logging/src/test/kotlin/eu/inqudium/limesium/reactive/logging/RequestLoggingWebFilterMetricsTest.kt
@@ -8,6 +8,7 @@ import ch.qos.logback.classic.turbo.TurboFilter
 import ch.qos.logback.core.read.ListAppender
 import ch.qos.logback.core.spi.FilterReply
 import eu.inqudium.limesium.common.BodyLogMode
+import eu.inqudium.limesium.common.EndpointLoggingMetrics
 import io.micrometer.core.instrument.Counter
 import io.micrometer.core.instrument.Gauge
 import io.micrometer.core.instrument.Meter